requirements.txt
pytest==7.4.3
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0